+++ vss_cli/data_types.py
@@ -48,6 +48,7 @@ class ConfigEndpoint:
     name: Optional[str] = None
     auth: Optional[str] = None
     token: Optional[str] = None
+    tf_enabled: Optional[bool] = False
 
     def __post_init__(self):
         """Post init method."""
@@ -76,7 +77,7 @@ class ConfigFile:
     )
 
     def get_endpoint(self, ep_name_or_url: str) -> List[ConfigEndpoint]:
-        """Get endpoind by name or url."""
+        """Get endpoint by name or url."""
         if self.endpoints:
             ep = list(
                 filter(lambda i: ep_name_or_url == i.name, self.endpoints)